PVOID is 64 bits when called from C# Structure size fixed

+++ HidLibraryCore/NativeMethods.cs
@@ -434,7 +434,7 @@ namespace HidLibrary
 	        public bool IsAlias;
 	        // 23 reserved bits
 	        [FieldOffset(16)]
-	        public UInt32  UserContext;
+	        public UInt64  UserContext;
         }
         
         [DllImport("hid.dll", CharSet = CharSet.Unicode)]
